requirements.txt
attrs==25.2.0
bcrypt==3.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.2
packaging==24.2
pycparser==2.22
pymongo==4.11.2
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3

# Testing dependencies
pytest==8.0.2
pytest-cov==4.1.0
pytest-docker==1.0.1
pytest-mock==3.12.0
coverage==7.4.3
